Add conversion webhook configuration to CAPI CRDs on startup
Since the CAPI CRD versions have been bumped to v1beta1, we need to provide a
way for the old objects with version v1alpha4 to be upgraded to the new version.
The convert webhook is already running in rancher-webhook, so the conversion
configuration is added to the CRDs on startup.

This conversion configuration cannot be added to the CRD ahead of time because
the CABundle is not known. We rely on the `needacert` controller to recognize
that the conversion configuration needs the CABundle, and it will be added when
the CRD is updated. If the conversion webhook configuration was added
beforehand, then Rancher startup would hang as it tries to watch the CAPI
objects. It would not be able to watch the CAPI objects because doing so would
call the conversion webhook which would fail because of an invalid cert (we
don't know the certificate information). Therefore, this migration is necessary.