Delete cleanup job if it times out
If the cleanup job timed out (like if the VM was deleted), then the node
will still exist in the cluster but it will have a Not Ready status.
Therefore, the cleanup job will get created to tolerate this status, but
the pod will never be scheduled because the VM was deleted.

After this change, if the cleanup job times out, then it will be
deleted.