Stop adding Rancher specific labels/annotations to CAPI machines
Rancher was adding labels and annotations to the CAPI machines to track things
like init nodes. The issue is that this caused the CAPI controllers to look at
the machines much more often than necessary. In addition, because it is
difficult for the Wrangler controllers to re-enqueue CAPI machines to the CAPI
controllers, it created timing/race conditions in certain scenarios. Finally,
the labels and annotations added by Rancher were for the dynamic planner in
Rancher, which is not CAPI related.

After this change, the Rancher-specific labels and annotations are moved to the
machine plan secret. This is a more appropriate place for these items because
CAPI doesn't know about or care about the machine plan secret.

This change also bumps lasso to fix a problem with the caches used by the CAPI
controllers. This change allows us to remove the machinestatus controller
because it is no longer necessary. (It was used as a way to re-enqueue CAPI
machines to the CAPI controllers, but the bump of lasso fixes the underlying
problem).

Lastly, common items needed by the various RKE2 controllers are moved to a
central location where they can all be referenced.