Delete oldest machines on scale down
By default, the CAPI controllers will delete a random machine when a machine
deployment is scaled down. This is different from the RKE1 implementation, which
 deletes the oldest machine first.

After this change, a deletion policy will be added to the machine deployments
with a deletion policy to delete the oldest machine on scale down.