Re-enqueue infra machine if provider ID set with jobComplete false
There is a possible race that could cause the jobComplete field to be false
(indicating that the creation job has not finished successfully), but the
providerID field set (indicating that the machine is running and has been
bootstrapped). These are contradictory states.

This change will re-enqueue the job if the infrastructure machine gets into this
state, which will set the jobComplete field and re-enqueue the infrastructure
machine to be processed again.

In addition, it is possible for the `GetJobName` to return a different
value every time if the infrastructure machine name is too long. This
change mitigates that by using labels to get the job.