Disallow deleting cloud credential referenced by provisioning cluster
Currently, it is possible to delete a cloud credential if it is referenced by a
provisioning cluster. This causes issues if a user wants to add or delete nodes
after the cloud credential is gone.

After this change, Rancher will prevent a user from deleting a cloud credential
if it is referenced by a provisioning cluster.