Allow node driver download in air-gapped environments
The harvester node driver is bundled in Rancher in order to provide support for
air-gapped installs of Rancher to provision Harvester clusters. However, the
download URL passed to the create job was still the external URL. This wouldn't
allow Harvester clusters to be created in air-gapped environments.

After this change, any enabled node driver will be copied to the appropriate
location so that the create job can download it directly from Rancher.