Send Rancher cert to rancher-machine
A previous change allowed rancher-machine to get the driver from Rancher
itself (as opposed to getting it from GitHub). This made air-gapped
environments possible for Harvester. However, this caused an issue if
Rancher was running with a self-signed cert.

This change sends the Rancher cert to the rancher-machine pod so that
the cert is trusted.