Fix group handling for impersonation
When CRTBs and PRTBs are created, they are either created for Users or
Groups, not both. Impersonation only needs to be set up for a user, and
the impersonation account must be able to impersonate the groups that
are found in the user's attributes. This change ensures that we don't
try to create an impersonation account for a group and removes unneeded
code for when an account is being created for a user.