Pass Rancher certs to rancher-machine job
In order for Harvester provisioning to work in air-gap, the driver needs to be
downloaded from the Rancher server. In addition, if the Rancher server is using
a self-signed certificate, that certificate needs to be passed to the Rancher
job so it can trust it.

This change constructs a secret with the Rancher certificate information and
passes it to the rancher-machine job.