Fix secretmigrator assemblers and s3 snapshots
An early draft of the secretmigrator controller used resource-specific
keys when creating the secrets, and when that was refactored the
assemblers were never updated, so for example the S3 secret assembler
was looking for a key 'secretKey' in the secret data, when it was
actually stored under the 'credential' key.

This also fixes the etcdbackup controller to use the assembled
BackupConfig, instead of the BackupConfig from the newly fetched Cluster
which does not have the credential on it.