Add spec.upgrade.envFrom to SUC upgrade watch
When Rancher is migrated, a new service-account token is created for
system-agent. However, this new service-account was not being deployed to the
downstream nodes because SUC wouldn't detect that the change occurred.

After this change, SUC will be told to watch spec.upgrade.envFrom for changes
and deploy a new plan when this field changes. This field is the correct one for
SUC to watch because the secret name contains a hash of the service-account
token that needs to be updated.