Wait for version upgrades before proceeding with more upgrades
Currently, when upgrading an RKE2 cluster, the planner will go onto the next
machine type before the upgrade finishes. For example, the planner will move
onto upgrade worker nodes before the control plane nodes are completely
upgraded. That is because the plan was considered in-sync after the system-agent
applied it and not after the upgrade completed.

Now, a plan will remain in-sync at the point the system-agent has applied it,
but the machine will not be considered to be in-sync until the version of the
node has been updated.

Because of this change, we must include the CAPI machines as "related resources"
so that the planner runs when the node version has been updated. This should
speed up provisioning slightly because the planner will kick off as soon as a
node has the ready state instead of waiting for other things to enqueue the
planner.