Use management context for catalog secrets
Without this patch, the secret assembler for Catalogs was trying to use
the secret lister from the UserContext which is not populated with
secrets from the management cluster. This was causing apps to not be
able to find the secret when deployed on downstream, non-local clusters.
This change passes the ScaledContext down to the helm controller so that
the secretmigrator assembler can use the correct secret lister.

(cherry picked from commit 10b81273928f89c913123002339b842fc56e2413)