Implement configurable lifecycle hook preStop sleep duration.
When used in combination with the aws-lb-controller (or any other
controller that implements a pod-readiness-gate) this will ensure that
any in-flight requests are correctly drained before the pod is
terminated.

The configured value should be greater than the LB deregistration delay value.

Example:

If you are using the default 300s deregistration delay for an AWS ALB/NLB,
then the preStop sleep duration value should be greater than 300.

Related issue:
https://github.com/kubernetes-sigs/aws-load-balancer-controller/issues/2366

Signed-off-by: Kris Gambirazzi <kris.gambirazzi@transferwise.com>