Use v2.6-head as default rancher-agent image
The master-head version of the rancher-agent is very old and does not support
Kubernetes v1.22+. This isn't a problem with "official" Rancher images because
the CATTLE_AGENT_IMAGE variable gets set to the proper values on build. However,
it does cause problems with local dev Rancher instances unless the
aforementioned environment variable is set.

This change uses the v2.6-head version of the agent to avoid problems in the
future.