Put statuses and errors on machine objects
In order to more clearly show a machine's state during upgrades and updates, the
planner will now use Provisioned and Updated conditions. Errors are also
detected when draining fails.

In addition, some logic was changed so that the planner will not move onto new
node types prematurely. For example, when the call was made to update the plan
secret, the `reconcile` would return if the plan secret was updated
successfully. The planner would take the return of a nil error as permission to
continue with the next node type.