Ensure machines have proper condition and labels
Custom Machines were not having their Ready condition marked as true even those
the status.ready field was marked true.

The CAPI machine controller expects the Ready condition on the infrastructure
machine to be true. Therefore, on Rancher upgrade, we will ensure this is the
case.

Additionally, custom machines were not given the CAPI cluster name
label. This change also ensures that:
1. All machines for each cluster are migrated, and
2. Each machine has both the CAPI and RKE2 cluster name labels