Stop labeling the control plane for designated init machines
By labeling the control plane when designating init machines, Rancher was unable
to continue with provisioning if there was a problem labeling the secret. For
example, if the bootstrap object was replaced for some reason after the machine
plan secret was labeled, then Rancher would not be able to label the new machine
plan secret and would be stuck.

Now, instead of checking the control plan object for the
`rke.cattle.io/init-node-machine-id-done` label, Rancher will check for the
init-node label on the secret.

Since this was the only place the `rke.cattle.io/init-node-machine-id-done` was
being used, it is being removed.