Display drain errors on machine (#36725)
* Put statuses and errors on machine objects

In order to more clearly show a machine's state during upgrades and updates, the
planner will now use Provisioned and Updated conditions. Errors are also
detected when draining fails.

In addition, some logic was changed so that the planner will not move onto new
node types prematurely. For example, when the call was made to update the plan
secret, the `reconcile` would return if the plan secret was updated
successfully. The planner would take the return of a nil error as permission to
continue with the next node type.

* Ensure machines have proper condition and labels

Custom Machines were not having their Ready condition marked as true even those
the status.ready field was marked true.

The CAPI machine controller expects the Ready condition on the infrastructure
machine to be true. Therefore, on Rancher upgrade, we will ensure this is the
case.

Additionally, custom machines were not given the CAPI cluster name
label. This change also ensures that:
1. All machines for each cluster are migrated, and
2. Each machine has both the CAPI and RKE2 cluster name labels

* Fix potential panic when control plane is nil on return

* Fix for custom RKE2 clusters and Ready condition

* Validate secret on retry

* Add more robust version checking

* Ensure plan annotations are as expected on retry