Use UserAttributes for impersonation
The impersonation module checks for Extra attributes in the user.Info
object and adds principal IDs and usernames from an auth provider to the
impersonation ClusterRole rules. Without this patch, this only works
when a user makes a request, because the request includes the
Extra data in the token, but does not work for the role template binding
controllers, which only send the user's UID and groups. The result is
that when a user sends a request, the ClusterRole is updated to include
the Extra data from the request, but after some time, the CRTB and PRTB
controllers will change it back.

With d9ba4d the user's principal IDs and usernames from the auth
provider should now be populated in the UserAttributes for that user, so
we can rely on that instead of whatever data may be in the user.Info
object. This way the data will be correct no matter whether this is
called from the request proxy or the role template binding controller.

If there is no auth provider, which is the case for system users, the
principal IDs and usernames are on the User object itself, and there is
no need to query UserAttributes.