Be more intelligent when the cluster-agent is disconnected
The fundamental issue being fixed here is restoring an RKE2 cluster with 1
etcd-only node, 1 control plane node, and 1 worker node. Previously, the etcd
and control plane nodes would be shutdown before the restore happened. The
cluster-agent would then be rescheduled to the worker node because it wouldn't
be updated. This caused some issues because it looked like the agent was
connected, but Rancher couldn't use it in any way because the api-service was
shutdown. This would cause the restore to hang because the etcd-only node's
kubelet version was not getting updated.

After this change, all nodes in the cluster are shutdown before the restore
happens. This caused the cluster agent to disconnect, as desired.

However, this caused another problem: since the CAPI clients and caches connect
through the cluster-agent, the caches would get out of sync and the kubelet
version was not being updated on the CAPI machines for a long time (several
minutes). This artificially slowed down the restore on all types of RKE2
clusters.

To fix this issue, a different client is passed to the CAPI cluster cache
reconciler that would cause the cache to be refreshed when the cluster-agent
disconnected.