Set a version in the default for the agent install scripts
If the setting value of the agent install scripts is different from the default,
then the code will go to the internet to try to download it. A previous change
set the value of the install script settings to a fixed version. However, the
default value remained at main. This caused Rancher to think that it needed to
download the script from the internet.

This change sets the default value to the expected value of the release so that
Rancher will not try to download the script.