Refactor node-cleanup job implementation
It was determined that some of the code around creating and deleting the
node-cleanup job was producing unexpected results like non-nil errors
with empty error messages. In addition, the cleanup job is being
leftover in some case.

This code makes the error messages more descriptive and adds retries to
try to ensure that the cleanup job gets deleted.