Add feature flag for custom RKE1 node cleanup
It is common for users to cleanup their own custom nodes when deleting
them from Rancher. This has caused issue with jobs and/or pods being
left around or nodes not being deleted properly.

This change adds a feature flag to turn the cleanup off globally if a
user doesn't need it.