Delete node-cleanup pods that aren't deleted
If a pod is stuck because it has been scheduled on a node that node
longer exists, then deleting the job may not be enough to delete this
pod. This would cause the job deletion to hang forever.

This change deletes the pod with gradePeriodSeconds of 0 to force its
deletion.