Pull Rancher images from private registry in RKE2 clusters
There were a few images like rancher-agent, system-agent, and
system-upgrade-controller that were not being pulled from the private registry
setup in an RKE2 cluster.

This change adds image resolvers, similar to those for v3 clusters, to ensure
that the images are being pulled from private registries whenever one is defined
for Rancher images.