Add better condition syncing for v2 provisioning objects
There were issues with different pages of the UI showing different status for
the cluster based on which object it was showing (management vs. provisioning).
This change adds better syncing between the cluster objects so they have the
same conditions and statuses.

In addition, the planner always used the Provisioned condition when
changing/applying the node plans. This was confusing, especially when a machine
was being updated. This change will use the Updated condition after the machine
is successfully provisioned the first time.