Continue removing imported cluster on error
Rancher will try to clean up imported clusters when they are removed. If this
failed for any reason, then the cluster would never get removed from Rancher.

This change adds a retry of 3 times: if Rancher is unable to clean up the
imported cluster, then it will move on with removal.