Configure controller option from env var
An environment variable, CATTLE_SKIP_NO_CHANGE_UPDATE, can now
be used to enable the lasso option of skipping enqueue for
update events that are not triggered by actual updates. This
mainly relates to resyncing controllers. Rancher has other
means of rerunning handlers that depend on other objects or
services and of retrying handlers that encounter errors or
must be run on an interval.