Stop creating ClusterAuthToken in kubeconfig API when hashing disabled
A previous change added token hashing implementation. When token hashing is
enabled, then ClusterAuthToken must be created in the downstream cluster when
the kubeconfig is generated. This means that the kubeconfig cannot be generated
until the cluster is active. If token hashing is disabled, then this requirement
is not needed and the kubeconfig can be generated without requiring that the
ClusterAuthToken be created in the downstream cluster.

This change will check the token hashing feature flag when generating a
kubeconfig and will not create the ClusterAuthToken. In addition, the
controller that would create the ClusterAuthToken in the downstream cluster is
restored to create the ClusterAuthToken if token hashing is disabled.