Handle IP addresses during RKE2 provisioning more robustly
There were several issues related to IP addresses that were causing issues with
RKE2 provisioning and multiple etcd-only nodes.

1. If there were multiple etcd-only nodes, then joinURL may change and be for
another node when the scraping happens. The reason was that the clientURLs would
be listed for each etcd node. When Rancher was scraping it, it was taking the
first one no matter what which may have been for another node.

The fix here is to add an addition periodic instruction to "cat" the db name
file and use it to identify the clientURL with the correct node.

2. The machine plan for the bootstrap node may have been created before the
machine had finished being created and the config saved to the machine state
secret. This would cause the plan to be missing IP addresses. Then, then next
plan would have the IP addresses, causing the cluster to look like it was
re-provisioning. If the user enabled the "drain on upgrade" feature, then the
nodes would all drain and their plan updated.

The fix here is to wait to assign a bootstrap node until we know that the
machine has finished creating, so that the machine state has the IP addresses,
if they are going to be available. Note that Rancher will not error if no IP
addresses are found because there may not be any depending on the provider.

3. Join URLs were being set in multiple places.

The fix here is to only set them in one place.

4. When getting the control plane join URL, Rancher was taking the URL from the
first control plane node that it found a join URL on. This could cause the join
URL to change after the node is bootstrapped.

The first control plane join URL is returned, regardless if it is set. This
ensures that join URL won't change after it is set on initial bootstrapping.