Remove probe-statuses when updating a machine plan
The system-agent will not set the probes as failed on the first failure
if the failure threshold is greater than 1. This is proper and expected.

However, if the plan is updated, we don't want to wait for the
failure-threshold. If the probes pass on the first try, then
system-agent will update them as such and Rancher will move on. However,
if the probes fail on the first try after the plan is updated, then
system-agent won't set the probe as successful and Rancher will wait for
successful probes before moving on.