Fixed snapshot RBAC for RKE2 clusters
Snapshots get stored in the local cluster under the `fleet-default`
namespace. Because of this, roles get access to individual
snapshots. If no snapshots are stored, the users do
not have access to view or create snapshots in the UI.
Users now have access to the `etcdsnapshots.rke.cattle.io`
in the namespace of the cluster. Making the CRD visible in
the schema when no snapshots are present.

The "Manage Cluster Backup" role also did not have the
`etcdsnapshots.rke.cattle.io` rule in its template.