[WIP] Optimize RTB lifecycle updates
When there are a large volume of projects and therefore a large volume
of role template bindings, running reflect.DeepEquals on controller
updates, even though there are only on the order of 5-10 rules in each
role template, starts to use a high proportion of memory. This change
introduces a different way of comparing the rules that shifts the
compute and memory burden toward the role template. The role template
controller now computes a checksum representing the rules and sets it as
an annotation on the resource. The RTB update then checks the checksum
to see whether it has changed since the last update, instead of checking
the rules directly.