[WIP] Look up access only for known schema
Without this, when a user creates a project, their total rolebindings
change and so their access changes. This means that the access set needs
to be rebuilt from scratch. If we know ahead of time what the user is
trying to do, we don't need to rebuild the entire access set, we can
just build the the set relevant to the requested access.