Added TTL for all kubeconfig tokens
Created a new setting 'kubeconfig-default-token-ttl-minutes'
for managing the TTL of kubeconfig tokens. This setting applies
to all kubeconfig tokens returned to users except for tokens
returned through the CLI. TTL for tokens created by the CLI are set via
'kubeconfig-token-ttl-minutes'. 'kubeconfig-token-ttl-minutes' will
be deprecated and replaced by kubeconfig-default-token-ttl-minutes'
once the setting is removed.