Check impersonation role bindings in setup
Impersonation needs three resources: clusterroles, clusterrolebindings,
and serviceaccounts. Without this patch, we assume that if the
clusterrole exists, the clusterrolebinding also exists. In some cases,
if packets were dropped during creation or if the clusterrolebinding was
deliberately or accidentally deleted, the clusterrolebinding might not
exist and a proxy request would fail due to lack of permissions. This
patch ensures the setup is more resilient against missing resources.