Trigger user attribute refresh for empty extras
In d9ba4df ExtraByProvider was added to UserAttributes and would be
refreshed when a new login token was created. With 4fa4285 the
impersonation code began depending on ExtraByProvider existing in
UserAttributes. However, if the UserAttributes was created before
d9ba4df, it would not have ExtraByProvider unless a new token triggered
adding it, for example by logging out and logging back in. The effect is
that if a Rancher server was created prior to v2.6.2 (prior to d9ba4df)
and then upgraded to v2.6.5 (includes 4fa4285) then users would lose
access to their downstream clusters because the impersonation
clusterrole would be missing the provider extras. This change ensures
that if ExtraByProvider is empty, the refresh is triggered by a
controller instead of waiting for the user to get a new token.