Moved the the Kube API provisiong package outside of rancher, but kept it in the clients package. This way integration tests can use the provisioning Kube API provisioning client if needed.
Updated the generated management clients. Updated the hosted provider extensions to account for the new management client updates.

Implemented a new schema for the provisioing.cattle.io API, and also updated the codegen package to generate the provisioing.cattle.io client, as well as the management.cattle.io clients.

Updated tests and packages to account for the new provisioning.cattle.io client.

Updated the rancher client to account for the new provisioning client. Also added helper to reinstantiate a client for when permissions change for a user. Added a helper that instantiates a kube api provisioning client.

Fixed Custom cluster tests.

Updated nodes configuration to allow for custom cluster tests have a way to configure the path to ssh path.