Move cluster and catalog secrets to Spec
Norman assumes that no updates to Status happen through the API,
including through the store, and so overwrites any changes[1]. This
means it is impossible to set a new value on Status for clusters or
catalogs upon update, it can only be done on create. This was resulting
in references to secrets being lost.

This change addresses the issue by moving secret values from Status to
Spec for clusters and catalogs.

Implementation notes:
- Notifiers are unaffected because the secret reference is the only
  value in status, since there are no status values in the original
  object it is not overwritten.
- SourceCodeProviderConfigs are unaffected because the secret reference
  was not in status.
- ClusterTemplateRevisions are unaffected because they are never
  updated, only new versions are created.
- The service account token is never updated through the norman store
  and so is unaffected.
- To avoid a complicated migration, unaffected resources are left alone
  even though it introduces inconsistency between the different types of
  resources.
- A getter is added for clusters and catalogs to check both status and
  spec for the secret values. The alternative is to add a migration in a
  controller.

[1] https://github.com/rancher/norman/blob/e21cabdc65b9f556c5cba1a5f2c49297850bdc46/types/convert/merge/merge.go#L13-L15