Expose rancher version endpoint
Expose the rancher version as an unauthenticated endpoint on
/rancherversion. The version and git commit are set at build time, but
the "prime" flag is set at runtime by an environment variable in the
Dockerfile. The endpoint is /rancherversion instead of /version because
the /version endpoint is just passed through to kubernetes and gives the
local cluster kubernetes version. The endpoint is unauthenticated rather
than authenticated because the k8s /version is already exposed[1] as
well as the /v1-{k3s,rke2}-release/* endpoints, so security through
obscurity doesn't buy us much here.

[1] https://github.com/kubernetes/kubernetes/issues/84040