Fix secrets migrator
Previously, there were a few problems with how objects were being
passed and used. Conditions were being set on clusters, but the
change was not commited with a client Update call. Now, condition
changes are followed by an update.

A Do function would pass in one pointer but then return a
different pointer value pointing to a more updated cluster object.
The code would then proceed to use the outdated object. This would
cause the service account token secret migration update to fail.
This would cause the service account token secret migration update
to fail. This was not a problem for some types of clusters like
AKSV2 clusters because there is redundant code for migrating the
service account token in their respective sync handlers and the
migration code in clusters.go would short circuit before the
problematic code was hit. Now, the returned object and object that
was passed in are set to have the same pointer value which is up
to date.

Modifications were made to the object referenced by the pointer
passed into the sync function. This is bad since this object will
be passed to subsequent handlers in the case that a nil cluster
object is returned. Now, the originally referenced object is not
modified.