Refresh ExtrasByProvider from derived tokens
If a user has no fresh UI login tokens, they may still need to log in
using a kubeconfig or API token. Without this change, if the provider
refresher happens to have run, ExtraByProvider will become empty, which
will cause requests to fail because the impersonation logic depends on
extras found in UserAttributes. This change ensures that extras are
refreshed either from login tokens or derived tokens, as long as the
user still has access to the provider.