Updating build process to use Go 1.19
Go 1.17, which has been used, is no longer supported by the Go
community. This update brings the build process up to the latest
version of Go.

golangci-lint is updated because version 1.44 did not work with Go
1.19 due to changes.

A large number of files were updated because gofmt handles comments
differently as of Go 1.18. There is now a markdown-ish format.
gofmt -ew . was run on the codebase to fix the formatting. This
primairly updated code comments to be properly formatted. In some
cases it fixed indentation elsewhere in a Go file.

Note: the zypper command was removed because it is a no-op in
practice as the latest Go is already in the image with that tag.

Signed-off-by: Matt Farina <matt.farina@suse.com>