Updating go.mod to Go 1.19
Previous commits update to use Go 1.19 for builing Rancher. This
updates Go modules to be 1.19 based to match the Go version of
Go being used.

Note, the zypper command to install 1.17 was previously an operation
that installed the 1.17 version of Go but did not use it. This
version was installed alongside 1.19 and not used in the path.

Signed-off-by: Matt Farina <matt.farina@suse.com>