Ensure cattle token secret has unique name
The "cattle" service account on a downstream cluster is the account that
Rancher uses to connect as an admin to the downstream cluster. Without
this change, if the cattle service account's token is deleted, the
cluster agent will regenerate it identically. This is a problem because
it makes rotation of the token nontrivial.

We can't craft the JWT ourselves or influence what claims are included
in it, that is done within kubernetes. The only way to change the
resulting JWT is to change the values kubernetes uses for claims. The
only option is to make the secret name unique[1]. All other claims come
from the service account, which we do not want to have to change in
order to rotate the token.

This change addresses the problem by using GenerateName when creating
the secret so that it will be unique every time. However, since the name
is no longer predictable, this causes problems when Rancher tries to
look up the token. We now need to look up the name of the secret from
the service account object. A further complication is that for
kubernetes 1.24, the secret name is no longer stored on the service
account, so now we set it explicitly. An extra benefit of this approach
is that we no longer create multiple tokens for service accounts on k8s
<1.24, since creating the token is skipped if it is found on the service
account.

This change refactors any code that was creating a service account token
to use the serviceaccounttoken.EnsureSecretForServiceAccount function in
order to be consistent everywhere. The function is updated to use a
backoff routine instead of an infinite loop to check the state of the
secret. It is flexible enough to use controller caches for callers with
access to that, and falls back to regular clients for remote callers
such as the agent.

See also the change that introduced this functionality in Rancher[2].

[1] https://github.com/kubernetes/kubernetes/blob/v1.25.2/pkg/serviceaccount/legacy.go#L39
[2] https://github.com/rancher/rancher/pull/38113

(cherry picked from commit 37c7ee194ab63dd0c98c4ce7b7411faa88f525ef)
(cherry picked from commit f08c50fa8ea207cbd310e67c8af8cf3b6d13df72)