Fix secretmigrator conditions and copies
Make the setting of conditions consistent. Conditions do not need to be
set explicitly when DoUntilTrue is used, but the object returned from
DoUntilTrue does need to be explicitly updated. Use copies of the object
until the object is ready to be updated.

(cherry picked from commit dd1f35b2dcde242280680b99c162be0a2489549f)