Make secretmigrator assemblers safe
The function comments on the secretmigrator assembler functions
indicated they would never change the original object, but this was not
true. This change ensures the objects are deepcopied within the function
so make it consistent with their documentation. It also removes
now-unnecessary deepcopies from the calling functions. Also corrects a
badly copy-pasted comment on one assembler function.