Set source to empty from "env"
Prior, when a setting was configured using an env var of the format
CATTLE_<setting-name>, the source was set to "env". That source
would then persist even when the env var was unset in the future.
This prevented users from ever changing a setting value through the
rancher API or UI. Now, unsetting the env var during an upgrade
will cause the source to be reset. This will allow users to change
the value again as long as the setting is no longer longer set by
an env var at setup.