Support empty string env var
Prior, if an empty string was set using an env var, getting the
setting from rancher api would return the empty value. However,
the value in ETCD would not be set. Once settings are interacted
with using anything besides norman, this will result in a change
of behavior where the value before the one last set is returned.
This is because norman is setting the value to be empty string.
Now, when the value is set to empty string, it will be set to
empty string in ETCD. This is valid since an empty string
indicates that the client should use the default value instead.