Update cluster service account token on change
Previously, if the cattle token was rotated on an RKE2 downstream, the
token would not propagate to the local cluster. This was not an issue
for RKE clusters because the service account token is updated through
the clusterdeploy controller and kontainer-engine driver. RKE2 clusters
only ever had the token set in the tunnel authorizer, and only upon
cluster creation. This change updates the tunnel authorizer to compare
and update the token if necessary. The local secret is fully rotated
rather than updated in-place in order to ensure the user context
controller is triggered so that user controllers are automatically
refreshed.